requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP & Networking
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Storage & File Handling
supabase==2.3.0
pillow==10.1.0
python-magic==0.4.27

# Payment Processing
stripe==7.8.0

# Communications
twilio==8.11.0

# Caching & Session Management
redis==5.0.1
aioredis==2.0.1

# Background Jobs & Scheduling
celery==5.3.4
celery[redis]==5.3.4
python-crontab==3.0.0

# Data Processing & Validation
validators==0.22.0
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development & Testing (optional in production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Audio Processing
pydub==0.25.1

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Webhook verification
hmac

# Image processing
Pillow==10.1.0

# Email
emails==0.6.0

# Async utilities
anyio==4.1.0